(HR) Attrition Report
A Human Resources (HR) Attrition Report in Power BI is a visual and analytical representation of employee turnover within an organization. 
Attrition refers to the rate at which employees leave a company, whether voluntarily (resignation, retirement) or involuntarily (termination, layoff). 
This type of report provides valuable insights to HR professionals, managers, and executives to understand the factors contributing to attrition, 
identify trends, and make informed decisions to improve employee retention.